+++ NEWS.md
@@ -15,6 +15,7 @@
 - Publication frequency introduced [MODLD-817](https://folio-org.atlassian.net/browse/MODLD-817)
 - PublicationFrequencyDictionary introduced [MODLD-826](https://folio-org.atlassian.net/browse/MODLD-826)
 - IllustrationDictionary introduced [MODLD-841](https://folio-org.atlassian.net/browse/MODLD-841)
+- Update IRIs with 'marc' to use 'library' instead [MODLD-792](https://folio-org.atlassian.net/browse/MODLD-792)
 
 ## 1.0.1 (03-12-2025)
 - Initial release